Refactor configuration management by replacing config.js with configService.js for improved environment variable handling
