+++ backend/src/index.js
@@ -1,5 +1,5 @@
 import app from './app.js';
-import config from './config/config.js';
+import config from './services/configService.js';
 
 const PORT = config.port;
